Made something runnable
Made changes to the board to include a function to get the move letters,
and fixed the constructor and class structure, and added a __deepcopy__
method.

Made changes to the search functions and heuristics and made the data
for the tree node the index of the move, its letter prior to making the
move, and the puzzle after making the move.

Got something to run, but the results are not the expected results.